requirements.txt
--editable ./lib/ns_workflow # ns-workflow==0.1
coverage>=5.1
cryptography>=2.9.2
Django
django-bootstrap4
django-cleanup
django-crispy-forms>=1.9.1
django-nose>=1.4.6
fabric-comanage-api
kafka-python==1.4.7
ldap3==2.7
mozilla-django-oidc>=1.2.3
ns-jwt>=0.1.2
psycopg2-binary>=2.8.6
pycryptodome>=3.9.7
PyJWT>=1.7.1
pytest-cov>=2.9.0
pytest-django>=3.9.0
pytest>=5.4.3
python-dateutil>=2.8.1
python-dotenv>=0.13.0
pytz>=2020.1
requests-oauthlib>=1.3.0
simplejson>=3.17.0
uWSGI
